backend/requirements.txt
psycopg[binary]>=3.1
SQLAlchemy>=2.0
fastapi>=0.110
uvicorn>=0.27
hypercorn>=0.16
pydantic-settings>=2.2
python-multipart
python-dotenv>=1.0
alembic>=1.13
PyJWT>=2.8.0
httpx>=0.27.0
prometheus-client>=0.20.0